pyproject.toml
[tool.poetry]
name = "improved-auto-conjugate-gradien-attack-develop"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">= 3.11,<3.12"
numpy = "^1.24.3"
colorlog = "^6.7.0"
PyYAML = "^6.0"
Cython = "^0.29.35"
easydict = "^1.10"
robustbench = {git = "https://github.com/RobustBench/robustbench.git"}
torch = {url = "https://download.pytorch.org/whl/cu118/torch-2.0.1%2Bcu118-cp311-cp311-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu118/torchvision-0.15.2%2Bcu118-cp311-cp311-linux_x86_64.whl"}

[tool.poetry.group.dev.dependencies]
matplotlib = "^3.7.1"
black = "^23.3.0"
pylint = "^2.17.4"
mypy = "^1.3.0"
GitPython = "^3.1.31"
openpyxl = "^3.1.2"
plotly = "^5.14.1"
jupyter = "^1.0.0"
notebook = "^6.5.4"
latex = "^0.7.0"
seaborn = "^0.13.0"
kaleido = "0.2.1"
optuna = "^3.4.0"
scikit-learn = "^1.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
